refactor(reconcile): extract shared reconcile-on-read helper
read-path commands (slice list, milestone list, project get) each
duplicated ~20 lines of STATE.md reconcile logic. extract to
application/reconcile/reconcile-on-read.ts, which encapsulates
active-milestone lookup, render closure construction, and error
swallowing. read commands now call it in one line.

no behavior change: read commands still succeed on reconcile failure,
still no-op when no open milestone exists.
