+++ src/cli/commands/project-get.cmd.ts
@@ -1,7 +1,5 @@
-import { join } from "node:path";
 import { getProject } from "../../application/project/get-project.js";
-import { reconcileState } from "../../application/reconcile/reconcile-state.js";
-import { renderStateMd } from "../../application/sync/generate-state.js";
+import { reconcileOnRead } from "../../application/reconcile/reconcile-on-read.js";
 import { isOk } from "../../domain/result.js";
 import { createClosableStateStoresUnchecked } from "../../infrastructure/adapters/sqlite/create-state-stores.js";
 import type { CommandSchema } from "../utils/flag-parser.js";
@@ -35,28 +33,7 @@ export const projectGetCmd = async (args: string[]): Promise<string> => {
 		createClosableStateStoresUnchecked();
 	const result = await getProject({ projectStore });
 
-	// Reconcile STATE.md against DB-derived render. Non-fatal: never blocks the read.
-	try {
-		const activeMR = milestoneStore.listMilestones();
-		if (activeMR.ok) {
-			const active = activeMR.data.find((m) => m.status !== "closed");
-			if (active) {
-				await reconcileState({
-					stateMdPath: join(process.cwd(), ".tff-cc", "STATE.md"),
-					renderStateMd: async () => {
-						const r = renderStateMd(
-							{ milestoneId: active.id },
-							{ milestoneStore, sliceStore, taskStore },
-						);
-						if (!r.ok) throw new Error(r.error.message);
-						return r.data;
-					},
-				});
-			}
-		}
-	} catch {
-		// Intentionally swallowed: reconcile must never fail a read.
-	}
+	await reconcileOnRead(process.cwd(), { milestoneStore, sliceStore, taskStore });
 
 	if (isOk(result)) {
 		if (result.data === null) {
